go/consensus/tendermint: Move full/seed services into own packages
