+++ go/consensus/tendermint/full/light.go
@@ -1,4 +1,4 @@
-package tendermint
+package full
 
 import (
 	"context"
@@ -20,11 +20,7 @@ func init() {
 }
 
 // Implements LightClientBackend.
-func (t *tendermintService) GetSignedHeader(ctx context.Context, height int64) (*consensusAPI.SignedHeader, error) {
-	if !t.features.Has(consensusAPI.FeatureServices) {
-		return nil, consensusAPI.ErrUnsupported
-	}
-
+func (t *fullService) GetSignedHeader(ctx context.Context, height int64) (*consensusAPI.SignedHeader, error) {
 	if err := t.ensureStarted(ctx); err != nil {
 		return nil, err
 	}
@@ -45,11 +41,7 @@ func (t *tendermintService) GetSignedHeader(ctx context.Context, height int64) (
 }
 
 // Implements LightClientBackend.
-func (t *tendermintService) GetValidatorSet(ctx context.Context, height int64) (*consensusAPI.ValidatorSet, error) {
-	if !t.features.Has(consensusAPI.FeatureServices) {
-		return nil, consensusAPI.ErrUnsupported
-	}
-
+func (t *fullService) GetValidatorSet(ctx context.Context, height int64) (*consensusAPI.ValidatorSet, error) {
 	if err := t.ensureStarted(ctx); err != nil {
 		return nil, err
 	}
@@ -67,11 +59,7 @@ func (t *tendermintService) GetValidatorSet(ctx context.Context, height int64) (
 }
 
 // Implements LightClientBackend.
-func (t *tendermintService) GetParameters(ctx context.Context, height int64) (*consensusAPI.Parameters, error) {
-	if !t.features.Has(consensusAPI.FeatureServices) {
-		return nil, consensusAPI.ErrUnsupported
-	}
-
+func (t *fullService) GetParameters(ctx context.Context, height int64) (*consensusAPI.Parameters, error) {
 	if err := t.ensureStarted(ctx); err != nil {
 		return nil, err
 	}
@@ -88,9 +76,6 @@ func (t *tendermintService) GetParameters(ctx context.Context, height int64) (*c
 }
 
 // Implements LightClientBackend.
-func (t *tendermintService) State() syncer.ReadSyncer {
-	if t.mux == nil {
-		return nil
-	}
+func (t *fullService) State() syncer.ReadSyncer {
 	return t.mux.State().Storage()
 }
