+++ go/oasis-node/cmd/debug/dumpdb/dumpdb.go
@@ -16,7 +16,6 @@ import (
 	beacon "github.com/oasisprotocol/oasis-core/go/beacon/api"
 	"github.com/oasisprotocol/oasis-core/go/common/logging"
 	consensus "github.com/oasisprotocol/oasis-core/go/consensus/genesis"
-	tendermint "github.com/oasisprotocol/oasis-core/go/consensus/tendermint"
 	"github.com/oasisprotocol/oasis-core/go/consensus/tendermint/abci"
 	abciState "github.com/oasisprotocol/oasis-core/go/consensus/tendermint/abci/state"
 	tendermintAPI "github.com/oasisprotocol/oasis-core/go/consensus/tendermint/api"
@@ -26,6 +25,7 @@ import (
 	roothashApp "github.com/oasisprotocol/oasis-core/go/consensus/tendermint/apps/roothash"
 	schedulerApp "github.com/oasisprotocol/oasis-core/go/consensus/tendermint/apps/scheduler"
 	stakingApp "github.com/oasisprotocol/oasis-core/go/consensus/tendermint/apps/staking"
+	tendermintCommon "github.com/oasisprotocol/oasis-core/go/consensus/tendermint/common"
 	epochtime "github.com/oasisprotocol/oasis-core/go/epochtime/api"
 	genesis "github.com/oasisprotocol/oasis-core/go/genesis/api"
 	genesisFile "github.com/oasisprotocol/oasis-core/go/genesis/file"
@@ -104,7 +104,7 @@ func doDumpDB(cmd *cobra.Command, args []string) {
 	ldb, _, stateRoot, err := abci.InitStateStorage(
 		ctx,
 		&abci.ApplicationConfig{
-			DataDir:           filepath.Join(dataDir, tendermint.StateDir),
+			DataDir:           filepath.Join(dataDir, tendermintCommon.StateDir),
 			StorageBackend:    storageDB.BackendNameBadgerDB, // No other backend for now.
 			MemoryOnlyStorage: false,
 			ReadOnlyStorage:   viper.GetBool(cfgDumpReadOnlyDB),
